+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed permissions after moving gitlab CI files
 - fixed linux build script syntax
 - REMIX-3048: Added slider widget for material attribute displace_out
+- Added PyCharm Debugger extension
 
 ### Changed
 - Updated hdremix to dd92d0f